Flask/requirements.txt
Flask
boto3
numpy
tensorflow
python-dotenv
